+++ R/api.R
@@ -1,6 +1,6 @@
 #' Set the required trakt.tv API credentials
 #'
-#' `get_trakt_credentials` searches for your credentials and stores them
+#' `trakt_credentials` searches for your credentials and stores them
 #' in the appropriate `option` variables.
 #' It also sets the HTTP header required for v2 API calls.
 #' To make this work automatically, place your key as environment variables in
@@ -28,15 +28,15 @@
 #' \dontrun{
 #' # Use a values set in ~/.Renviron in an R session:
 #' # (This is automatically executed when calling library(tRakt))
-#' get_trakt_credentials()
+#' trakt_credentials()
 #'
 #' # Explicitly set values in an R session, overriding .Renviron values if present
-#' get_trakt_credentials(
+#' trakt_credentials(
 #'   username = "sean",
 #'   client.id = "12fc1de7671c7f2fb4a8ac08ba7c9f45b447f4d5bad5e11e3490823d629afdf2"
 #' )
 #' }
-get_trakt_credentials <- function(username, client.id,
+trakt_credentials <- function(username, client.id,
                                   silent = TRUE) {
   username <- ifelse(missing(username), Sys.getenv("trakt_username"), username)
   client_id <- ifelse(missing(client.id), Sys.getenv("trakt_client_id"), client.id)
@@ -66,7 +66,7 @@ get_trakt_credentials <- function(username, client.id,
 #' `trakt.api.call` makes an API call to a specified URL and returns the parsed output.
 #'
 #' @param url APIv2 method. See \href{http://docs.trakt.apiary.io/}{the trakt API}.
-#' @param client.id API client id. see [get_trakt_credentials] for further information.
+#' @param client.id API client id. see [trakt_credentials] for further information.
 #' @param convert.datetime If `TRUE` (default), known top-level datetime variables
 #' are converted to `POSIXct`. This might miss some variables and does not recurse
 #' into nested lists or list-columns.