+++ R/zzz.R
@@ -1,7 +1,7 @@
 # nocov start
 
 .onLoad <- function(...) {
-  get_trakt_credentials()
+  trakt_credentials()
 }
 
 # nocov end