+++ tests/testthat/test-api.R
@@ -2,17 +2,17 @@ context("test-test-api")
 
 test_that("Client ID is set without .Renviron", {
   client.id <- "12fc1de7671c7f2fb4a8ac08ba7c9f45b447f4d5bad5e11e3490823d629afdf2"
-  expect_message(get_trakt_credentials(silent = FALSE))
+  expect_message(trakt_credentials(silent = FALSE))
 
   Sys.setenv(trakt_client_id = "")
-  expect_failure(expect_message(get_trakt_credentials()))
+  expect_failure(expect_message(trakt_credentials()))
   expect_equal(getOption("trakt.client.id"), client.id)
 
   expect_message(
-    get_trakt_credentials(username = "arbitraryusername", silent = FALSE)
+    trakt_credentials(username = "arbitraryusername", silent = FALSE)
   )
   expect_equal(getOption("trakt.username"), "arbitraryusername")
-  expect_message(get_trakt_credentials(client.id = client.id, silent = FALSE))
+  expect_message(trakt_credentials(client.id = client.id, silent = FALSE))
 })
 
 test_that("trakt.api.call can make API calls", {